[outputs] refs #47 - Move outputs' models to models folder and consequent problems solved.
