+++ main.go
@@ -5,7 +5,6 @@ import (
 
 	_ "github.com/fibercrypto/FiberCryptoWallet/src/coin/skycoin"
 	_ "github.com/fibercrypto/FiberCryptoWallet/src/models"
-	_ "github.com/fibercrypto/FiberCryptoWallet/src/models/outputs"
 	"github.com/therecipe/qt/core"
 	"github.com/therecipe/qt/gui"
 	"github.com/therecipe/qt/qml"
